Working in a concept of caching values.  The trick here, is that for records like attack, they are just identifiers to other records.  For example, an Attack has reference to a unit, and an ability. This is important for the calculation engine, except that the calculation engine cannot do much with just the identifier.
Rather than have the engine have to manage copies of each record, and look them up, it made sense to me to expose the Attacker for example, as a get accessor.

The get accessor will then use the identifier stored in teh Attack, to lookup a Unit in the Unit Cache, which is just a hash table.

The unit cache in this case, is managed when units are first created.

Anyways, it's a work in progress...